@@ -11,7 +11,7 @@
     "start": "gatsby develop -H 0.0.0.0",
     "build": "gatsby build",
     "build:profile": "gatsby build --profile",
-    "serve": "gatsby serve",
+    "serve": "gatsby serve -H 0.0.0.0",
     "lint": "pnpm run lint:js && pnpm run lint:css",
     "lint:js": "eslint .",
     "lint:css": "stylelint 'src/**/*.css'",
@@ -19,7 +19,7 @@
     "eslint:check": "eslint-config-prettier .eslintrc.json",
     "fonts:preload-cache": "gatsby-preload-fonts"
   },
-  "keywords": ["website", "gatsby", "react"],
+  "keywords": ["website", "gatsby", "react", "tailwindcss"],
   "author": "Mohamed Elkholy <mohatt@pm.me>",
   "license": "MIT",
   "repository": {
@@ -31,45 +31,42 @@
   "dependencies": {
     "@mdx-js/mdx": "1.6.22",
     "@mdx-js/react": "1.6.22",
-    "@tailwindcss/typography": "0.4.1",
-    "autoprefixer": "10.2.6",
+    "autoprefixer": "10.3.1",
     "classnames": "2.3.1",
-    "gatsby": "3.7.2",
+    "gatsby": "3.10.2",
     "gatsby-plugin-advanced-pages": "2.2.1",
-    "gatsby-plugin-catch-links": "3.7.1",
-    "gatsby-plugin-graphql-config": "0.16.1",
-    "gatsby-plugin-manifest": "3.7.1",
-    "gatsby-plugin-mdx": "2.7.1",
-    "gatsby-plugin-postbuild": "3.0.1",
-    "gatsby-plugin-postcss": "4.7.1",
-    "gatsby-plugin-preload-fonts": "2.7.1",
-    "gatsby-plugin-react-helmet": "4.7.1",
-    "gatsby-plugin-sharp": "3.7.1",
-    "gatsby-plugin-sitemap": "4.3.1",
-    "gatsby-remark-images": "5.4.1",
-    "gatsby-remark-prismjs": "5.4.1",
-    "gatsby-remark-smartypants": "4.4.1",
-    "gatsby-source-filesystem": "3.7.1",
-    "gatsby-transformer-sharp": "3.7.1",
-    "gatsby-transformer-yaml": "3.7.1",
+    "gatsby-plugin-graphql-config": "0.19.0",
+    "gatsby-plugin-image": "1.10.1",
+    "gatsby-plugin-manifest": "3.10.0",
+    "gatsby-plugin-mdx": "2.10.1",
+    "gatsby-plugin-postbuild": "3.0.2",
+    "gatsby-plugin-postcss": "4.10.0",
+    "gatsby-plugin-preload-fonts": "2.10.0",
+    "gatsby-plugin-react-helmet": "4.10.0",
+    "gatsby-plugin-sharp": "3.10.2",
+    "gatsby-plugin-sitemap": "4.6.0",
+    "gatsby-remark-images": "5.7.0",
+    "gatsby-remark-smartypants": "4.7.0",
+    "gatsby-source-filesystem": "3.10.0",
+    "gatsby-transformer-sharp": "3.10.0",
+    "gatsby-transformer-yaml": "3.10.0",
     "lodash": "4.17.21",
-    "moment": "2.29.1",
-    "postcss": "8.3.5",
+    "postcss": "8.3.6",
     "postcss-import": "14.0.2",
     "postcss-preset-env": "6.7.0",
     "postcss-url": "10.1.3",
-    "prismjs": "1.23.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-helmet": "6.1.0",
-    "tailwindcss": "2.2.2"
+    "react-image-lightbox": "5.1.4",
+    "tailwindcss": "2.2.7"
   },
   "devDependencies": {
     "@mohatt/eslint-config-gatsby-standard": "3.2.0",
-    "@types/tailwindcss": "2.0.6",
-    "eslint": "7.29.0",
-    "prettier": "2.3.1",
-    "simple-icons": "5.2.0",
+    "@types/tailwindcss": "2.2.1",
+    "eslint": "7.31.0",
+    "prettier": "2.3.2",
+    "simple-icons": "5.8.0",
     "stylelint": "13.13.1",
     "stylelint-config-standard": "22.0.0",
     "webpack-bundle-analyzer": "4.4.2"
@@ -77,12 +74,6 @@
   "engines": {
     "node": ">=12.17.0",
     "pnpm": ">=6"
-  },
-  "pnpm": {
-    "overrides": {
-      "gatsby>postcss": "8.3.5"
-    },
-    "neverBuiltDependencies": ["fsevents"]
   },
   "browserslist": [
     "last 2 versions",
