@@ -1,36 +1,36 @@
-import React, { useContext, useEffect, useRef, useState } from 'react'
+import React, { useEffect, useRef, useState } from 'react'
 import classNames from 'classnames'
 import { generatePath, routeExists } from 'gatsby-plugin-advanced-pages'
-import { useCurrentPath, useSiteMetadata } from '../../hooks'
+import { site, $window, $document } from '../../constants'
+import { usePath } from '../../hooks'
 import { Link } from '..'
-import { LayoutContext } from './Layout'
 
-function MenuItemSubs({ items, onClick, hashPath, currentPath }) {
-  const hashElms = useRef([])
+function SubMenu({ items, hashPath, currentPath, onClick }) {
+  const hashTargets = useRef([])
   const [activeHash, setActiveHash] = useState(null)
 
   useEffect(() => {
-    hashElms.current = items.map(i => i.hash && document.getElementById(i.hash))
-    const elms = hashElms.current.filter(Boolean)
-    if (elms.length === 0) {
+    hashTargets.current = items.map(i => i.hash && $document.getElementById(i.hash))
+    const targets = hashTargets.current.filter(Boolean)
+    if (targets.length === 0) {
       setActiveHash(null)
       return
     }
 
     let busy = false
-    const scrollListner = function () {
+    const listener = () => {
       if (!busy) {
         busy = true
-        requestAnimationFrame(function () {
-          const { innerHeight } = window
+        requestAnimationFrame(() => {
+          const { innerHeight } = $window
           // Minimum intersecting area of the element to be considered active (40% vh)
           let max = innerHeight * 0.4
-          setActiveHash(elms.reduce(function (acc, el) {
-            const { top, height } = el.getBoundingClientRect()
+          setActiveHash(targets.reduce((acc, target) => {
+            const { top, height } = target.getBoundingClientRect()
             const intersect = Math.min(height, height + top, innerHeight - top)
             if (intersect >= max) {
               max = intersect
-              return el.id
+              return target.id
             }
             return acc
           }, null))
@@ -39,33 +39,31 @@ function MenuItemSubs({ items, onClick, hashPath, currentPath }) {
       }
     }
 
-    scrollListner()
-    window.addEventListener('scroll', scrollListner, { passive: true })
-    return function () {
-      window.removeEventListener('scroll', scrollListner)
-    }
-  }, [items, setActiveHash, currentPath])
+    listener()
+    $window.addEventListener('scroll', listener, { passive: true })
+    return () => $window.removeEventListener('scroll', listener)
+  }, [items, currentPath])
 
   return (
     <ul className='mr-2'>
-      {items.map(({ label, to, params, hash }, i) => (
+      {items.map(({ label, to, params, external, hash }, i) => (
         <li key={i}>
           <Link
             to={to || hashPath + '#' + hash}
-            params={params}
-            onClick={to || currentPath !== hashPath ? onClick : function (e) {
+            onClick={to || currentPath !== hashPath ? onClick : e => {
               onClick(e)
-              if (hashElms.current[i]) {
-                hashElms.current[i].scrollIntoView({ behavior: 'smooth', block: 'start' })
+              if (hashTargets.current[i]) {
+                hashTargets.current[i].scrollIntoView({ behavior: 'smooth', block: 'start' })
                 e.preventDefault()
               }
             }}
-            className={classNames(
-              'block mb-8 sm:mb-6 hover:text-primary',
-              hash && hash === activeHash && 'text-primary'
-            )}
-            activeClassName='text-primary'
+            className={classNames('block mb-8 sm:mb-6 hover:text-primary', hash && hash === activeHash && 'text-primary')}
             children={label}
+            {...external ? { external } : {
+              params: params,
+              activeClassName: 'text-primary',
+              partiallyActive: true,
+            }}
           />
         </li>
       ))}
@@ -73,43 +71,39 @@ function MenuItemSubs({ items, onClick, hashPath, currentPath }) {
   )
 }
 
-function MenuItem({ label, to, params, items, onClick }) {
-  const [, currentPath] = useCurrentPath()
-  let current = null
-  if (items) {
-    const href = routeExists(to)
-      ? generatePath(to, params, null, true)
-      : to
-    current = currentPath.startsWith(href) ? href : null
-  }
-
-  return (
-    <li>
-      <Link
-        to={to}
-        params={params}
-        onClick={onClick}
-        className='block mb-8 sm:mb-6 hover:text-typo'
-        activeClassName='text-typo active'
-        partiallyActive={to !== 'home'}
-        children={label}
-      />
-      {current && (
-        <MenuItemSubs items={items} hashPath={current} currentPath={currentPath} onClick={onClick} />
-      )}
-    </li>
-  )
-}
-
-function Menu({ isHome, className }) {
-  const { menu, title } = useSiteMetadata()
-  const { setMenuOpen } = useContext(LayoutContext)
-  const closeMenu = () => setMenuOpen(false)
+function Menu({ setMenuOpen, className }) {
+  const [currentPath] = usePath()
   return (
     <nav className={className}>
-      {isHome && <h1 className='hidden'>{title}</h1>}
       <ul className='font-display font-medium italic text-right mr-6'>
-        {menu.map((props, i) => <MenuItem key={i} onClick={closeMenu} {...props} />)}
+        {site.menu.map(({ label, to, params, external, items }, i) => {
+          let isActive = null
+          if (items && items.length) {
+            const href = routeExists(to)
+              ? generatePath(to, params, null, true)
+              : to
+            isActive = currentPath.startsWith(href) ? href : null
+          }
+
+          return (
+            <li key={i}>
+              <Link
+                to={to}
+                onClick={setMenuOpen}
+                className='block mb-8 sm:mb-6 hover:text-typo'
+                children={label}
+                {...external ? { external } : {
+                  params: params,
+                  activeClassName: 'text-typo active',
+                  partiallyActive: to !== 'home',
+                }}
+              />
+              {isActive && (
+                <SubMenu items={items} hashPath={isActive} currentPath={currentPath} onClick={setMenuOpen} />
+              )}
+            </li>
+          )
+        })}
       </ul>
     </nav>
   )
