@@ -1,7 +1,9 @@
+export { useLocation } from '@gatsbyjs/reach-router'
+
+export { usePath } from '../providers/path'
+export { useAnalytics, useAnalyticsCallback } from '../providers/analytics'
+export { useTheme } from '../providers/theme'
+
 export { default as useSiteMetadata } from './site-metadata'
 export { default as useLocalStorage } from './local-storage'
-export { useCurrentLocation, useCurrentPath } from './current-location'
-export { default as useCurrentRoute } from './current-route'
-export { useProjectSkills, useProjectSkill } from './project-skills'
-export { useProjectCategories, useProjectCategory } from './project-categories'
-export { useIsBrowser } from './util'
+export { useMounted } from './sensors'
