@@ -1,13 +1,16 @@
 import React from 'react'
+import { PathProvider } from './path'
 import { AnalyticsProvider } from './analytics'
 import { ThemeProvider } from './theme'
 
-export default function Providers({ children }) {
+export default function Providers({ children, location }) {
   return (
-    <AnalyticsProvider>
-      <ThemeProvider>
-        {children}
-      </ThemeProvider>
-    </AnalyticsProvider>
+    <PathProvider location={location}>
+      <AnalyticsProvider location={location}>
+        <ThemeProvider>
+          {children}
+        </ThemeProvider>
+      </AnalyticsProvider>
+    </PathProvider>
   )
 }
