@@ -1,14 +1,67 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { graphql } from 'gatsby'
 import classNames  from 'classnames'
-import { PlatformHandle } from '../commons/platforms'
-import { Page, Button, Heading, Link, Markdown, Section, Separator } from '../components'
+import { GatsbyImage } from 'gatsby-plugin-image'
+import Lightbox from 'react-image-lightbox'
+import { themeScreens } from '../constants'
+import { PlatformHandle } from '../util'
+import { Page, Heading, Link, Markdown, Section, Separator } from '../components'
 import { ProjectCategory, ProjectSkill, Testimonial } from './partials'
 
 const statuses = {
   CMP: 'Completed',
   ONG: 'Ongoing',
   ARC: 'Archived',
+}
+
+function Gallery({ screens }) {
+  const [index, setIndex] = useState(-1)
+
+  const size = screens.length
+  if (size === 0) {
+    return null
+  }
+
+  return (
+    <>
+      <div className='flex overflow-x-auto mb-12'>
+        {screens.map(({ thumb }, i) => {
+          const cls = `scr_thumb_${i}`
+          const ratio = thumb.width / thumb.height
+          return (
+            <React.Fragment key={i}>
+              <style>{`
+                .${cls} { width: ${200 * ratio}px }
+                @media(min-width: ${themeScreens.lg}) {
+                  .${cls} { width: ${300 * ratio}px }
+                }
+                @media(min-width: ${themeScreens['2xl']}) {
+                  .${cls} { width: ${500 * ratio}px }
+                }
+              `.replace(/\s+/g, '')
+              }</style>
+              <GatsbyImage
+                image={thumb}
+                className={`flex-shrink-0 cursor-zoom mr-1 ${cls}`}
+                onClick={() => setIndex(i)}
+                alt={`Screen ${i + 1}`}
+              />
+            </React.Fragment>
+          )
+        })}
+      </div>
+      {index > -1 && (
+        <Lightbox
+          mainSrc={screens[index].full.images.fallback.src}
+          nextSrc={screens[(index + 1) % size].full.images.fallback.src}
+          prevSrc={screens[(index + size - 1) % size].full.images.fallback.src}
+          onCloseRequest={() => setIndex(-1)}
+          onMovePrevRequest={() => setIndex((index + size - 1) % size)}
+          onMoveNextRequest={() => setIndex((index + 1) % size)}
+        />
+      )}
+    </>
+  )
 }
 
 function Metadata({ title, children, className }) {
@@ -30,65 +83,50 @@ export default class Project extends Page {
       id: pageContext.project,
     }
 
-    const image = project.image.childImageSharp.gatsbyImageData
+    const image = project.image.childImageSharp
     const screens = []
     if (project.hasImage) {
       screens.push(image)
-      this.image = image.images.fallback.src
+      this.image = image.thumb.images.fallback.src
     }
-    project.screens.forEach(s => s && screens.push(s.childImageSharp.gatsbyImageData))
+    project.screens.forEach(s => s && screens.push(s.childImageSharp))
 
     return (
       <Section id={project.categories[0]?.id}>
         <Heading title={this.title} primary>
           {this.description}
         </Heading>
-        {screens.length > 0 && (
-          <div className='flex overflow-x-auto mb-12'>
-            {screens.map((screen, i) => (
-              <img key={i} src={screen.images.fallback.src} alt={`Screen ${i + 1}`} />
-            ))}
-          </div>
-        )}
+        <Gallery screens={screens} />
         <div className='grid md:grid-cols-3 gap-x-4 gap-y-8 mb-12'>
           <Metadata title='Project Name'>{project.title}</Metadata>
           <Metadata title='Start Date'>{project.started}</Metadata>
           <Metadata title='Status'>{statuses[project.status] || statuses.CMP}</Metadata>
-          {project.skills.length > 0 && (
-            <Metadata title='Skills'>
-              <ProjectSkill.Map data={project.skills}>
-                {({ props }) => <Button color='alt' size='tiny' className='mr-1' {...props} />}
-              </ProjectSkill.Map>
-            </Metadata>
-          )}
-          {project.categories.length > 0 && (
-            <Metadata title='Categories'>
-              <ProjectCategory.Map data={project.categories}>
-                {({ props }) => <Button color='alt' size='tiny' className='mr-1' {...props} />}
-              </ProjectCategory.Map>
-            </Metadata>
-          )}
-          {project.handles.length > 0 && (
-            <Metadata title='Links'>
-              <PlatformHandle.Map data={project.handles}>
-                {({ title, href, Icon }) => (
-                  <Link to={href} external='project_link' className='inline-flex leading-5 mr-4 link'>
-                    <Icon className='h-5 mr-2' />
-                    {title}
-                  </Link>
-                )}
-              </PlatformHandle.Map>
-            </Metadata>
-          )}
+          <ProjectSkill.Map data={project.skills}>
+            {items => <Metadata title='Skills'>{items}</Metadata>}
+          </ProjectSkill.Map>
+          <ProjectCategory.Map data={project.categories}>
+            {items => <Metadata title='Categories'>{items}</Metadata>}
+          </ProjectCategory.Map>
+          <PlatformHandle.Map data={project.handles}>
+            {items => <Metadata title='Links'>{items}</Metadata>}
+            {({ title, href, Icon }) => (
+              <Link to={href} external='project_link' className='link mr-4'>
+                <Icon className='h-5 mr-2' />
+                <span>{title}</span>
+              </Link>
+            )}
+          </PlatformHandle.Map>
         </div>
         <div className='xl:max-w-3xl'>
-          {project.testimonials[0] && (
-            <div className='mb-12'>
-              <Separator />
-              <Testimonial test={project.testimonials[0]} className='text-primary my-8' />
-              <Separator />
-            </div>
-          )}
+          <Testimonial.Map data={project.testimonials} limit={1} className='text-primary my-8'>
+            {items => (
+              <div className='mb-12'>
+                <Separator />
+                {items}
+                <Separator />
+              </div>
+            )}
+          </Testimonial.Map>
           <Markdown>{project.body}</Markdown>
         </div>
       </Section>
@@ -107,12 +145,14 @@ export const query = graphql`
       hasImage
       image {
         childImageSharp {
-          gatsbyImageData(height: 500)
+          thumb: gatsbyImageData(height: 500, placeholder: BLURRED)
+          full: gatsbyImageData(layout: FULL_WIDTH, formats: [AUTO], breakpoints: [1920])
         }
       }
       screens {
         childImageSharp {
-          gatsbyImageData(height: 500)
+          thumb: gatsbyImageData(height: 500, placeholder: BLURRED)
+          full: gatsbyImageData(layout: FULL_WIDTH, formats: [AUTO], breakpoints: [1920])
         }
       }
       categories {
