@@ -1,84 +1,48 @@
 import React from 'react'
 import { graphql } from 'gatsby'
-import { useProjectSkills } from '../hooks'
+import { skillTags } from '../constants'
 import { Page, Heading, Icon, Link, Section } from '../components'
 import { ProjectSkill } from './partials'
 
-const tags = {
-  backend: [
-    { id: 'language', title: 'Languages', icon: 'code' },
-    { id: 'framework', title: 'Frameworks', icon: 'stack' },
-    { id: 'cms', title: 'CMSs', icon: 'tree' },
-    { id: 'database', title: 'Databases', icon: 'database' },
-    { id: 'api', title: 'APIs', icon: 'cloud' },
-    { id: 'test', title: 'Testing', icon: 'test' },
-  ],
-  frontend: [
-    { id: 'language', title: 'Languages', icon: 'code' },
-    { id: 'js', title: 'JavaScript', icon: 'javascript' },
-    { id: 'css', title: 'CSS', icon: 'css' },
-    { id: 'api', title: 'APIs', icon: 'cloud' },
-    { id: 'ssg', title: 'SSGs', icon: 'stack' },
-  ],
-  devops: [
-    { id: 'tool', title: 'Tools', icon: 'terminal' },
-    { id: 'deploy', title: 'Deployment', icon: 'cloudUp' },
-    { id: 'ci', title: 'CI/CD', icon: 'tools' },
-    { id: 'pkgman', title: 'Package Managers', icon: 'cloudDown' },
-    { id: 'git', title: 'Git Hosting', icon: 'gitRepo' },
-  ],
-  software: [
-    { id: 'dev', title: 'Development', icon: 'appStore' },
-    { id: 'comms', title: 'Communication', icon: 'discuss' },
-    { id: 'env', title: 'Environment', icon: 'computer' },
-  ],
+const skillTagGroups = {
+  back: ['lang', 'frame', 'cms', 'db', 'api', 'test'],
+  front: ['lang', 'frame', 'ui', 'api', 'ssg'],
+  devops: ['tool', 'deploy', 'ci', 'pkgm', 'git'],
+  soft: ['dev', 'env', 'comm'],
 }
 
-function SkillBlock({ icon, title, children }) {
+function SoftSkill({ icon, title, children }) {
   return (
-    <div className='flex text-lg leading-normal'>
-      <Icon name={icon} className='w-8 h-8 text-primary' />
-      <div className='flex-1'>
-        <h3 className='ml-2 leading-8 text-primary'>{title}</h3>
-        <div className='mt-4 -ml-3'>{children}</div>
-      </div>
+    <div className='text-lg'>
+      <h3 className='text-primary'>
+        <Icon name={icon} className='w-8 h-8 mr-2' />
+        <span>{title}</span>
+      </h3>
+      <div className='mt-3 ml-3'>{children}</div>
     </div>
   )
 }
 
-function SkillList({ tags, title, icon }) {
-  const skills = useProjectSkills(tags)
-  return skills.length > 0 && (
-    <div className='flex'>
-      <Icon name={icon} className='w-8 h-8 text-primary' />
-      <div className='flex-1'>
-        <h3 className='ml-2 leading-8 text-primary'>{title}</h3>
-        <ul className='mt-3 -ml-4 space-y-3 font-medium'>
-          <ProjectSkill.Map data={skills}>
-            {({ title, size, props, Icon }) => (
-              <li className='flex leading-6'>
-                {size > 0
-                  ? (
-                    <Link className='link' {...props}>
-                      <Icon className='h-6 mr-2' />{title} <sup>{size}</sup>
-                    </Link>
-                  )
-                  : <><Icon className='h-6 mr-2' />{title}</>
-                }
-              </li>
-            )}
-          </ProjectSkill.Map>
-        </ul>
-      </div>
-    </div>
-  )
-}
-
-function SkillTagGrid({ tag }) {
+function SkillTagGroup({ group, data }) {
   return (
-    <div className='grid xs:grid-cols-2 xs:gap-x-1 xl:grid-cols-3 xl:gap-x-4 gap-y-8 xl:max-w-4xl text-lg'>
-      {tags[tag].map(({ id, title, icon }) => (
-        <SkillList key={id} tags={[tag, id]} title={title} icon={icon} />
+    <div className='grid xs:grid-cols-2 xs:gap-x-1 md:grid-cols-3 md:gap-x-4 gap-y-8 xl:max-w-4xl text-lg font-medium leading-6'>
+      {skillTagGroups[group].map(tag => (
+        <ProjectSkill.Map key={tag} data={data} tags={[group, tag]} limit={6}>
+          {items => (
+            <div>
+              <h3 className='text-primary'>[{skillTags[tag]}]</h3>
+              <ul className='mt-3 ml-3 space-y-3'>{items}</ul>
+            </div>
+          )}
+          {({ title, size, props, Icon }) => (
+            <li>
+              {size > 0
+                ? <Link className='link' {...props}><Icon className='h-6 mr-2' /><span>{title} <sup>{size}</sup></span></Link>
+                : <><Icon className='h-6 mr-2' /><span>{title}</span></>
+              }
+            </li>
+          )}
+        </ProjectSkill.Map>
       ))}
     </div>
   )
@@ -86,28 +50,29 @@ function SkillTagGrid({ tag }) {
 
 export default class Skills extends Page {
   view() {
-    this.title = this.props.data.page.title
+    const { page: { title }, skills: { nodes } } = this.props.data
+    this.title = title
     return (
       <>
         <Section>
-          <Heading title={this.title} primary>
+          <Heading title={title} primary>
             Software development is hard. Managing projects is hard. Working remotely is hard. Here are some
             things I'm good at, to help ease the pain.
           </Heading>
           <div className='grid md:grid-cols-2 gap-8'>
-            <SkillBlock icon='bug' title='Problem Solving'>
+            <SoftSkill title='Problem Solving' icon='bug'>
               I can take vague problems and requirements and break them down into steps and solutions.
-            </SkillBlock>
-            <SkillBlock icon='server' title='Systems Thinking'>
+            </SoftSkill>
+            <SoftSkill title='Systems Thinking' icon='stack'>
               I'm good at thinking abstractly and putting together systems with many moving parts.
-            </SkillBlock>
-            <SkillBlock icon='discuss' title='Communicating'>
+            </SoftSkill>
+            <SoftSkill title='Communicating' icon='chat'>
               I can explain things clearly, communicate problems quickly and write accurately and concisely.
-            </SkillBlock>
-            <SkillBlock icon='calendar' title='Organising'>
+            </SoftSkill>
+            <SoftSkill title='Organising' icon='calendar'>
               I can self-manage, work to deadlines, organise projects and present well-structured
               and complete deliverables.
-            </SkillBlock>
+            </SoftSkill>
           </div>
         </Section>
         <Section id='backend'>
@@ -116,27 +81,27 @@ export default class Skills extends Page {
             back-end done by me. My main stack usually involves PHP with the CMS/Framework of
             choice, and alternatively Node.js.
           </Heading>
-          <SkillTagGrid tag='backend' />
+          <SkillTagGroup group='back' data={nodes} />
         </Section>
         <Section id='frontend'>
           <Heading title='Front-end Development'>
             I create responsive websites that allow the user to have the best and most appropriate
             experience suited to the device they are using.
           </Heading>
-          <SkillTagGrid tag='frontend' />
+          <SkillTagGroup group='front' data={nodes} />
         </Section>
         <Section id='devops'>
           <Heading title='DevOps'>
             I use these tools and cloud services to setup an integrated, effective and efficient
             workflows that meet the project needs.
           </Heading>
-          <SkillTagGrid tag='devops' />
+          <SkillTagGroup group='devops' data={nodes} />
         </Section>
         <Section id='software'>
           <Heading title='Software'>
             My local web development setup.
           </Heading>
-          <SkillTagGrid tag='software' />
+          <SkillTagGroup group='soft' data={nodes} />
         </Section>
       </>
     )
@@ -147,6 +112,14 @@ export const query = graphql`
   query Skills($id: String!) {
     page(id: { eq: $id }) {
       title
+    }
+
+    skills: allProjectSkill {
+      nodes {
+        ...ProjectSkillFragment
+        icon
+        tags
+      }
     }
   }
 `
