@@ -1,100 +1,94 @@
 import React from 'react'
 import { graphql } from 'gatsby'
-import { PlatformHandle } from '../../commons/platforms'
-import { Button, Link } from '../../components'
+import { GatsbyImage, getImage } from 'gatsby-plugin-image'
+import { PlatformHandle } from '../../util'
+import { Button, Link, Pagination } from '../../components'
 import { ProjectCategory, ProjectSkill } from '.'
 
-function filter(data, exclude, limit) {
-  const filtered = (exclude ? data.filter(o => o.id !== exclude) : data).slice(0, limit)
-  filtered.diff = data.length - filtered.length
-  return filtered
-}
-
-function ProjectCard({ project, children, skill, category }) {
-  const image = project.image.childImageSharp.gatsbyImageData
+function ProjectCard({ project, skill, category }) {
   const props = {
     to: 'projects.project',
     params: { project: project.slug },
-    title: 'View project details page',
-  }
-  const categories = filter(project.categories, category, 2)
-  const skills = filter(project.skills, skill, 4)
-
-  if (children instanceof Function) {
-    return children(Object.assign({}, project, { image, categories, skills, props }))
+    title: 'View project details',
   }
 
   return (
-    <div>
+    <div className='max-w-[430px]'>
       <div>
-        <div className='relative' style={{ maxWidth: image.width }}>
+        <div className='relative'>
           <Link className='block' {...props}>
-            <img className='w-full border-2 border-primary rounded-md shadow-lg' src={image.images.fallback.src} alt={project.title} />
+            <GatsbyImage
+              image={getImage(project.image)}
+              className='border-2 border-primary rounded-md shadow-lg'
+              alt={project.title}
+            />
           </Link>
-          {project.handles.length > 0 && (
-            <div className='absolute -bottom-4 right-4'>
-              <PlatformHandle.Map data={project.handles.slice(0, 2)}>
-                {({ title, href, Icon }) => (
-                  <Button
-                    to={href}
-                    external='project_card_link'
-                    title={title}
-                    size='mono'
-                    className='border-primary ml-3'
-                    children={<Icon className='h-5' />}
-                  />
-                )}
-              </PlatformHandle.Map>
-            </div>
-          )}
-          {categories.length > 0 && (
-            <div className='absolute -top-4 left-4'>
-              <ProjectCategory.Map data={categories}>
-                {({ props }) => <Button color='primary' size='tiny' className='mr-1' {...props} />}
-              </ProjectCategory.Map>
-              {categories.diff > 0 && (
-                <Button size='tiny' color='primary' {...props}>
-                  +{categories.diff}
-                </Button>
-              )}
-            </div>
-          )}
+          <PlatformHandle.Map data={project.handles} limit={2}>
+            {items => (
+              <div className='absolute -bottom-4 right-4'>
+                {items}
+              </div>
+            )}
+            {({ title, href, Icon }) => (
+              <Button
+                to={href}
+                external='project_card_link'
+                title={title}
+                size='mono'
+                className='!border-primary ml-3'
+                children={<Icon className='h-5' />}
+              />
+            )}
+          </PlatformHandle.Map>
+          <ProjectCategory.Map data={project.categories} exclude={category} limit={2} color='primary'>
+            {items => (
+              <div className='absolute -top-4 left-4'>
+                {items}
+              </div>
+            )}
+          </ProjectCategory.Map>
         </div>
       </div>
       <div className='mt-4'>
-        <h3>
-          <Link className='link-primary' {...props}>
-            {project.title}
-          </Link>
-        </h3>
-        {skills.length > 0 && (
-          <div className='mt-2'>
-            <ProjectSkill.Map data={skills}>
-              {({ props }) => <Button color='alt' size='tiny' className='mr-1' {...props} />}
-            </ProjectSkill.Map>
-            {skills.diff > 0 && (
-              <Button size='tiny' color='alt' {...props}>
-                +{skills.diff}
-              </Button>
-            )}
-          </div>
-        )}
-        <div className='mt-4'>
+        <Link className='link-primary' {...props}>
+          {project.title}
+        </Link>
+        <ProjectSkill.Map data={project.skills} exclude={skill} limit={8}>
+          {items => (
+            <div className='mt-3 flex overflow-hidden relative'>
+              {items}
+              <span className='overflow-grad h-full' />
+            </div>
+          )}
+        </ProjectSkill.Map>
+        <div className='mt-3 text-base max-h-12 overflow-hidden relative'>
           {project.desc}
+          <span className='overflow-grad h-1/2' />
         </div>
       </div>
     </div>
   )
 }
 
-ProjectCard.Map = function ProjectCardMap({ data, children, ...props }) {
+ProjectCard.Map = function ProjectCardMap({ data, ...props }) {
   return (data.edges || data).map(project => {
     if (data.edges) {
       project = project.node
     }
 
-    return <ProjectCard key={project.slug} project={project} children={children} {...props} />
+    return <ProjectCard key={project.slug} project={project} {...props} />
   })
+}
+
+ProjectCard.Grid = function ProjectCardGrid({ data, paginated, params }) {
+  return (
+    <div className='grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-x-4 gap-y-8 text-lg'>
+      <ProjectCard.Map data={data} category={params.category} skill={params.skill} />
+      {paginated && (
+        <Pagination route={paginated} params={params} pageInfo={data.pageInfo} />
+      )}
+    </div>
+  )
 }
 
 export const ProjectCardFragment = graphql`
@@ -104,7 +98,7 @@ export const ProjectCardFragment = graphql`
     desc
     image {
       childImageSharp {
-        gatsbyImageData(aspectRatio: 1.8, width: 430)
+        gatsbyImageData(aspectRatio: 1.8, width: 430, placeholder: BLURRED)
       }
     }
     handles
@@ -113,6 +107,19 @@ export const ProjectCardFragment = graphql`
     }
     skills {
       ...ProjectSkillFragment
+    }
+  }
+`
+
+export const ProjectCardGridPaginatedFragment = graphql`
+  fragment ProjectCardGridPaginatedFragment on ProjectConnection {
+    edges {
+      node {
+        ...ProjectCardFragment
+      }
+    }
+    pageInfo {
+      ...Pagination
     }
   }
 `
