@@ -1,24 +1,29 @@
 import React from 'react'
 import { graphql } from 'gatsby'
+import { createReactMap } from '../../util'
+import { Button } from '../../components'
 
-function ProjectCategory({ category, children }) {
-  if(!category.props) {
-    category.props = {
+function ProjectCategory({ category, children, color = 'alt' }) {
+  let { id, title, size, props } = category
+  if(!props) {
+    props = category.props = !size ? { children: title } : {
       to: 'projects.category',
-      params: { category: category.id },
-      children: category.title,
-      title: `View ${category.size} project${category.size !== 1 ? 's' : ''} published under "${category.id}" category`,
+      params: { category: id },
+      children: title,
+      title: `View ${size} project${size !== 1 ? 's' : ''} published under "${id}" category`,
     }
   }
 
-  return children(category)
+  return children
+    ? children(category)
+    : <Button color={color} size='tiny' className='mr-1' {...props} />
 }
 
-ProjectCategory.Map = function ProjectCategoryMap({ data, children }) {
-  return data.map(category => (
-    <ProjectCategory key={category.id} category={category} children={children} />
-  ))
-}
+ProjectCategory.Map = createReactMap(function ProjectCategoryMap(category, { exclude, color, children }){
+  return exclude !== category.id && (
+    <ProjectCategory key={category.id} category={category} color={color} children={children} />
+  )
+})
 
 export const ProjectCategoryFragment = graphql`
   fragment ProjectCategoryFragment on ProjectCategory {
